Adds support for Bamboo 2.0.0
Bamboo changed the API signature for bamboo 2.0.0 for adapters.

deliver/2 must now return an :ok or :error tuple. Since we are
delegating to the next adapter and only updating the config, we can
leave that change to the downstream adapter.

The other change is that the adapter must implement
supports_attachments?/0. Since we delegate, we don't know if we support
attachments. Since the function must be there and we have no access to
the data to say if attachments are supported, I've decided always to
return true.

Closes #4